requirements.txt
# Consolidated requirements for Generative Operator Analysis Framework

# ============================================================================
# Core Scientific Computing
# ============================================================================
numpy>=1.21.0
scipy>=1.7.0
pandas>=2.1.4

# ============================================================================
# Machine Learning and Statistics
# ============================================================================
scikit-learn>=1.0.0
statsmodels>=0.13.0

# ============================================================================
# Deep Learning
# ============================================================================
torch>=1.11.0
torchvision>=0.12.0

# ============================================================================
# Visualization
# ============================================================================
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# ============================================================================
# Image Processing
# ============================================================================
scikit-image>=0.19.0
Pillow>=8.0.0

# ============================================================================
# Data Formats and I/O
# ============================================================================
h5py>=3.1.0
netCDF4>=1.5.0
zarr>=2.10.0

# ============================================================================
# Progress Bars and Utilities
# ============================================================================
tqdm>=4.62.0
click>=8.0.0
watchdog>=2.1.0

# ============================================================================
# Logging and Monitoring
# ============================================================================
tensorboard>=2.8.0
wandb>=0.12.0
structlog>=23.2.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation>=0.42b0

# ============================================================================
# Performance Monitoring and Profiling
# ============================================================================
psutil>=5.8.0
memory-profiler>=0.60.0
line-profiler>=3.5.0
py-spy>=0.3.0
snakeviz>=2.1.0
prometheus-client>=0.19.0
prometheus-api-client>=0.5.3
grafana-api>=1.0.3

# ============================================================================
# Configuration and Serialization
# ============================================================================
pyyaml>=6.0
python-dotenv>=1.0.0
hydra-core>=1.3.2
toml>=0.10.0
jsonschema>=4.0.0
orjson>=3.9.10
msgpack>=1.0.7
cloudpickle>=3.0.0

# ============================================================================
# Testing Frameworks
# ============================================================================
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=2.5.0
pytest-timeout>=2.1.0
pytest-benchmark>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
coverage>=6.0.0
codecov>=2.1.0

# ============================================================================
# Code Quality and Formatting
# ============================================================================
black>=22.0.0
isort>=5.10.0
flake8>=5.0.0
mypy>=0.990
pylint>=2.15.0
pre-commit>=2.20.0

# ============================================================================
# Security
# ============================================================================
bandit>=1.7.0
safety>=2.0.0
cryptography>=41.0.7
jwt>=1.3.1
bcrypt>=4.1.2

# ============================================================================
# Documentation
# ============================================================================
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
sphinx-autodoc-typehints>=1.19.0
myst-parser>=0.18.0
nbsphinx>=0.8.0

# ============================================================================
# Jupyter Support
# ============================================================================
jupyter>=1.0.0
jupyterlab>=3.4.0
notebook>=6.4.0
ipython>=8.0.0
ipywidgets>=7.6.0

# ============================================================================
# Parallel Processing
# ============================================================================
joblib>=1.1.0
dask>=2022.0.0
celery>=5.3.4
kombu>=5.3.4

# ============================================================================
# Web Framework and API
# ============================================================================
fastapi[all]>=0.85.0
uvicorn[standard]>=0.18.0
starlette>=0.27.0
pydantic>=2.5.0
httpx>=0.23.0
aiohttp>=3.9.1
requests>=2.31.0
urllib3>=2.0.7
gunicorn>=21.2.0

# ============================================================================
# Database Support
# ============================================================================
sqlalchemy>=2.0.23
pymongo>=4.0.0
asyncpg>=0.29.0
redis>=5.0.1
alembic>=1.8.0

# ============================================================================
# Build and Packaging
# ============================================================================
build>=0.8.0
twine>=4.0.0
wheel>=0.37.0
setuptools>=65.0.0
setuptools-scm>=7.0.0

# ============================================================================
# Development Tools
# ============================================================================
gitpython>=3.1.0
factory-boy>=3.2.0
responses>=0.21.0
freezegun>=1.2.0
faker>=15.0.0

# ============================================================================
# Cloud and Container Support
# ============================================================================
kubernetes>=28.1.0
docker>=6.1.3

# ============================================================================
# CI/CD Utilities
# ============================================================================
tox>=3.25.0
nox>=2022.8.7

# ============================================================================
# Production Deployment
# ============================================================================
supervisor>=4.2.5

# ============================================================================
# Performance and Load Testing
# ============================================================================
locust>=2.12.0

# ============================================================================
# GPU Monitoring (Optional)
# ============================================================================
nvidia-ml-py3>=7.352.0

# ============================================================================
# Time Series and Analytics
# ============================================================================
influxdb-client>=1.38.0

# ============================================================================
# Message Queue Support (Optional)
# ============================================================================
asyncio-mqtt>=0.13.0

# ============================================================================
# Platform-specific (Optional)
# ============================================================================
# systemd-python>=235  # Linux systems only, uncomment if needed
# terraform>=1.6.0  # Via subprocess, not pip
# helm>=3.13.0  # Via subprocess, not pip

# ============================================================================
# Optional ML/AI Acceleration (uncomment if needed)
# ============================================================================
# triton>=2.1.0  # NVIDIA Triton
# onnxruntime>=1.16.3  # ONNX Runtime
# tensorrt>=8.6.1  # TensorRT (NVIDIA)

# ============================================================================
# Optional Message Queues (uncomment if needed)
# ============================================================================
# pika>=1.3.2  # RabbitMQ
# kafka-python>=2.0.2  # Apache Kafka
